+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v6.29](https://github.com/replicant4j/replicant/tree/v6.29) (2019-02-12)
 [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.28...v6.29)